+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1-alpha.0](https://github.com/dockite/dockite/compare/@dockite/core@0.3.0...@dockite/core@0.3.1-alpha.0) (2020-06-26)
+
+**Note:** Version bump only for package @dockite/core
+
+
+
+
+
 # [0.3.0](https://github.com/dockite/dockite/compare/@dockite/core@0.2.2...@dockite/core@0.3.0) (2020-06-25)